🐛 Fix Windows portability and docstring lint in the ABC bridge
Two test assumptions did not hold on Windows:

- `os.access(path, os.X_OK)` is true for any existing file there, since Windows
  has no execute bit, so the non-executable rejection cannot be exercised.
- `shutil.which()` resolves through `PATHEXT`, so an extension-less stub named
  `berkeley-abc` is never found. The stub now gets a `.exe` suffix on Windows.

Also move the named script wrappers into their own module: `__init__.py` is
meant to hold only the docstring and re-exports, and the per-file exemption for
that rule is specific to the top-level package.

The attribute docstrings on module constants tripped `check-docstring-first`,
which reads them as second module docstrings; they are plain comments now.

Finally, exempt the bridge from DOC502. Its public functions deliberately
document the exceptions that propagate out of their helpers -- dropping those
entries would leave callers unaware that, say, `resyn2` can raise
`AbcNotFoundError`.

These were missed locally because `prek` only inspects git-tracked files and the
bridge was still untracked when lint last ran.
