+++ pyproject.toml
@@ -195,8 +195,9 @@ isort.required-imports = ["from __future__ import annotations"]
 "docs/**" = ["T20", "ERA001", "INP001"]
 "noxfile.py" = ["T20", "TID251", "ERA001"]
 "tools/*.py" = ["S603", "S607"]
-# the ABC bridge shells out to a user-configured executable by design
-"python/aigverse/abc/*.py" = ["S603"]
+# the ABC bridge shells out to a user-configured executable by design, and its
+# public functions document the exceptions that propagate out of their helpers
+"python/aigverse/abc/*.py" = ["S603", "DOC502"]
 "*.pyi" = ["D418", "PYI002", "PYI021", "A003", "N999"]
 "*.ipynb" = [
     "D", # pydocstyle
