+++ python/aigverse/abc/__init__.py
@@ -19,8 +19,6 @@ Example:
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 from ._binary import (
     ABC_ENV_VAR,
     abc_binary,
@@ -30,11 +28,18 @@ from ._binary import (
     set_abc_binary,
 )
 from ._errors import AbcError, AbcExecutionError, AbcNotFoundError, AbcTimeoutError
-from ._runner import AigT, run_commands, run_script
+from ._runner import run_commands, run_script
 from ._scripts import SCRIPTS, expand_script
-
-if TYPE_CHECKING:
-    import os
+from ._wrappers import (
+    compress,
+    compress2,
+    compress2rs,
+    dc2,
+    resyn,
+    resyn2,
+    resyn2rs,
+    resyn3,
+)
 
 __all__ = [
     "ABC_ENV_VAR",
@@ -60,259 +65,3 @@ __all__ = [
     "run_script",
     "set_abc_binary",
 ]
-
-
-def resyn(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``resyn`` script on a network.
-
-    A short balance/rewrite loop; the lightest of the standard scripts.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("resyn"), timeout=timeout, verbose=verbose, binary=binary)
-
-
-def resyn2(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``resyn2`` script on a network.
-
-    The most widely used ABC size-reduction script.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("resyn2"), timeout=timeout, verbose=verbose, binary=binary)
-
-
-def resyn3(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``resyn3`` script on a network.
-
-    A resubstitution-driven variant of ``resyn``.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("resyn3"), timeout=timeout, verbose=verbose, binary=binary)
-
-
-def compress(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``compress`` script on a network.
-
-    Like ``resyn``, but every command runs in level-preserving mode.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("compress"), timeout=timeout, verbose=verbose, binary=binary)
-
-
-def compress2(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``compress2`` script on a network.
-
-    Like ``resyn2``, but every command runs in level-preserving mode.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("compress2"), timeout=timeout, verbose=verbose, binary=binary)
-
-
-def resyn2rs(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``resyn2rs`` script on a network.
-
-    ``resyn2`` extended with resubstitution passes; slower and usually smaller.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("resyn2rs"), timeout=timeout, verbose=verbose, binary=binary)
-
-
-def compress2rs(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``compress2rs`` script on a network.
-
-    ``resyn2rs`` in level-preserving mode; the strongest of these scripts.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("compress2rs"), timeout=timeout, verbose=verbose, binary=binary)
-
-
-def dc2(
-    ntk: AigT,
-    *,
-    timeout: float | None = None,
-    verbose: bool = False,
-    binary: str | os.PathLike[str] | None = None,
-) -> AigT:
-    """Runs ABC's ``dc2`` script on a network.
-
-    ABC's builtin combinational don't-care-based optimization.
-
-    The script is expanded into builtin ABC commands, so it does not depend on
-    an ``abc.rc`` being present. See :data:`SCRIPTS` for the exact expansion.
-
-    Args:
-        ntk: The combinational network to optimize.
-        timeout: Seconds to wait for ABC to terminate, or ``None`` for no limit.
-        verbose: If ``True``, print everything ABC wrote.
-        binary: Overrides the resolved ABC executable for this call only.
-
-    Returns:
-        The optimized network, of the same type as ``ntk``.
-
-    Raises:
-        TypeError: If ``ntk`` is a ``SequentialAig`` or not an ``Aig`` at all.
-        AbcNotFoundError: If no ABC executable could be located.
-        AbcTimeoutError: If ABC did not terminate within ``timeout`` seconds.
-        AbcExecutionError: If ABC reported an error or produced no usable output.
-    """
-    return run_script(ntk, expand_script("dc2"), timeout=timeout, verbose=verbose, binary=binary)
